wp12_hackathon/from_pdf_to_figure_prototype/requirements.txt
PyMuPDF>=1.23.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0